requirements.txt
docker
setuptools
fastapi
uvicorn[standard]
sqlmodel
asyncpg
alembic
pydantic-settings
email-validator
# Тесты
pytest
pytest-asyncio
httpx